% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS.R
\name{CheckFastQC}
\alias{CheckFastQC}
\title{verify if FASTQC is accessible by the library
path location should be saved in the ConfigFile. Try first CheckFastQC().}
\usage{
CheckFastQC()
}
\value{
see details
}
\description{
verify if FASTQC is accessible by the library
path location should be saved in the ConfigFile. Try first CheckFastQC().
}
\details{
it should output the help page of FASTQC
}
\author{
Elmer A. Fernández
}
