% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralRAPIFunctions.R
\name{.GetMHCgeneList}
\alias{.GetMHCgeneList}
\title{.GetMHCgeneList}
\usage{
.GetMHCgeneList(
  segmentsLength = 10,
  alleleGroup = c("HLA", "BoLA", "Gogo", "H", "H2", "Mamu", "Patr", "SLA", "All")
)
}
\arguments{
\item{segmentsLength}{number of segments}

\item{alleleGroup}{allele group}
}
\value{

}
\description{
generate a ggplot2, where in the y axis the amount of alleles bindig to the sequence starting at x position is displayed
internal
}
\examples{
\dontrun{
FormatOut(eDB = result)
}
}
