% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS.R
\name{.OpenConfigFile}
\alias{.OpenConfigFile}
\title{internal function to load configuration info.}
\usage{
.OpenConfigFile()
}
\value{
the configuration list. Each slot holds one software
}
\description{
internal function to load configuration info.
}
\details{
It will create a RAPI_Confing.RDS file with software location paths and version
for NGS data processing. FastQC, TrimGalore,
 PLEASE DO NOT DELET this file
}
