% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS.R
\name{.SaveConfigFile}
\alias{.SaveConfigFile}
\title{internal function to save configuration info.}
\usage{
.SaveConfigFile(conf)
}
\description{
internal function to save configuration info.
}
\details{
It will save the RAPI_Confing.RDS file after installation of TrimGalore, FASTQC
}
