% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralRAPIFunctions.R
\name{FormatOut}
\alias{FormatOut}
\title{FormatOut}
\usage{
FormatOut(resRAPIMHC)
}
\arguments{
\item{resRAPIMHC}{an object results}
}
\value{

}
\description{
Internal function, it format the output. This function should not be used
}
\examples{
\dontrun{
FormatOut(eDB = result)
}
}
