% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralRAPIFunctions.R
\name{GetAlleleList}
\alias{GetAlleleList}
\title{GetMHCgeneList}
\usage{
GetAlleleList(
  nSegments = 10,
  MHC.group = c("HLA", "BoLA", "Gogo", "H", "H2", "Mamu", "Patr", "SLA", "DRB", "DQ",
    "DP", "Mouse", "All"),
  MHCII = FALSE
)
}
\arguments{
\item{nSegments}{integer, number of segments}

\item{MHC.group}{MHC allele group "HLA","BoLA", "Gogo","H","H2", "Mamu", "Patr","SLA",
"DRB", "DQ","DP","Mouse","All"}

\item{MHCII}{(logical) if MHCII = TRUE, then only MHC.group in "DRB", "DQ","DP","Mouse","All" are evaluated
if MHCII = FALSE, only "HLA","BoLA", "Gogo","H","H2", "Mamu", "Patr","SLA", "All" are evaluated}
}
\value{
a list of alleles with length = nSegments (ussualy set to the number of cores or CPU workers)
}
\description{
returns an list of Allele types (length nSegments)
}
\examples{
\dontrun{
FormatOut(eDB = result)
}
}
