% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRAPIFunctions.R
\name{ImmunoDominantRegions}
\alias{ImmunoDominantRegions}
\title{ImmunoDominantRegions}
\usage{
ImmunoDominantRegions(seq, eDB, pr = c("inverse", "count"), nCores)
}
\arguments{
\item{seq}{a sequence from seqinr::read.fasta}

\item{eDB}{a data frame or RAPIMHCI or RAPIMHCII class object}

\item{pr}{The rank level. (0.5 by default)if not missing, the level of selected top rank. if pr <= 0 it just counts the AA presence in a peptide candidate}

\item{nCores}{(integer) number of cores to use, if missing it will be automatically set}
}
\value{
a ggplot2 object
}
\description{
generate a ggplot2 logo of the evaluated papetide and its binding regions (Core) allowing
the identification of the Core with its deleted (bulging out) regions and predicted insertion sites
}
\examples{
\dontrun{
seq <- seqinr::read.fata(file = ........)
ImmunoDominantRegions(seq = seq, eDB = res)
}
}
