% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS.R
\name{InstallFastQC}
\alias{InstallFastQC}
\title{painless FastQC installation.
please provide the url from https://www.bioinformatics.babraham.ac.uk/projects/download.html#fastqc}
\usage{
InstallFastQC(fqurl)
}
\arguments{
\item{fqurl}{the url of fastqc}
}
\value{
if success, the configuration list file will be printed and the approrpiate fastqc version will appears
}
\description{
painless FastQC installation.
please provide the url from https://www.bioinformatics.babraham.ac.uk/projects/download.html#fastqc
}
\details{
It will install FastQC, the RAPI_Config.RDS file will be updated in accordance
}
\author{
Elmer A. Fernández
}
