% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS.R
\name{InstallTrimGalore}
\alias{InstallTrimGalore}
\title{painless trim_galore installation.
please provide the url from https://github.com/FelixKrueger/TrimGalore}
\usage{
InstallTrimGalore(trimTarBall, where, cutadaptPath)
}
\arguments{
\item{trimTarBall}{the url of trimgalore, like "https://github.com/FelixKrueger/TrimGalore/archive/0.6.5.tar.gz"}

\item{where}{(ptional) directory to where the software will be installed. if missing, it will use ~/.local hidden directory}

\item{cutadaptPath}{the diretory where cutadapt was installed, if missing it assumes ~/.local/bin/cutadapt}
}
\value{
if success, the TrimGalore run without inputs and cutadapt location
}
\description{
painless trim_galore installation.
please provide the url from https://github.com/FelixKrueger/TrimGalore
}
\details{
It will install TrimGalore, the RAPI_Config.RDS file will be updated in accordance
}
\author{
Elmer A. Fernández
}
