% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRAPIFunctions.R
\name{PeptideBindersLogo}
\alias{PeptideBindersLogo}
\title{PeptideBindersLogo}
\usage{
PeptideBindersLogo(eDB, size = c(8:14), type = c("Peptide", "InterCore"))
}
\arguments{
\item{eDB}{a data frame or RAPIMHCI or RAPIMHCII class object}

\item{size}{a integer or integer verctor with values between 8 and 14}

\item{type}{character Peptide or InterCore for RAPIMHC or Peptide for RAPIMHCII}
}
\value{
a ggplot2 object
}
\description{
generate a ggplot2 logo of all peptides binder of lenth in size.
if eDB is an RAPIMHCII object only one plot will be generated
}
\examples{
\dontrun{
RepresentedRegions(resDF)
}
}
