% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRAPIFunctions.R
\name{PeptideCoreLogo}
\alias{PeptideCoreLogo}
\title{PeptideBindersLogo}
\usage{
PeptideCoreLogo(eDB, peptide)
}
\arguments{
\item{eDB}{a data frame or RAPIMHCI or RAPIMHCII class object}

\item{peptide}{a string sequence peptide}
}
\value{
a ggplot2 object
}
\description{
generate a ggplot2 logo of all cores binders associated to such peptide.
The cores is the sequence predicted as binder inside the peptide, it can be the same sequence or
a shorter one (a deletion/loop in the tested sequence) or longer one with an insertion.
}
\examples{
\dontrun{
PeptideBindersLogo(eDB = res, peptide == "MSRIGMEVTPSGTW")
}
}
