% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRAPIFunctions.R
\name{PlotPeptideCore}
\alias{PlotPeptideCore}
\title{PlotPeptideCore}
\usage{
PlotPeptideCore(eDB, index)
}
\arguments{
\item{eDB}{a data frame or RAPIMHCI or RAPIMHCII class object}

\item{index}{(integer) indicating which row of the RAPIMHC and RAPIMHCII objects}
}
\value{
a ggplot2 object
}
\description{
generate a ggplot2 logo of the evaluated papetide and its binding regions (Core) allowing
the identification of the Core with its deleted (bulging out) regions and predicted insertion sites
}
\examples{
\dontrun{
PlotPeptideCore(eDB = res, index = 1)
}
}
