% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRAPIFunctions.R
\name{PlotPeptideLengthDistribution}
\alias{PlotPeptideLengthDistribution}
\title{PlotPeptideLengthDistribution}
\usage{
PlotPeptideLengthDistribution(resDF, main)
}
\arguments{
\item{resDF}{a data frame or RAPIMHCI or RAPIMHCII class object}

\item{main}{a cg¿haracter with a plot title}
}
\value{
a ggplot2 object
}
\description{
generate a ggplot2 bar plot of the distribution counts of predicted peptides of different length
}
\examples{
\dontrun{
RepresentedRegions(resDF)
}
}
