% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRAPIFunctions.R
\name{RepresentedRegions}
\alias{RepresentedRegions}
\title{PlotPeptideLengthDistribution}
\usage{
RepresentedRegions(
  resDF,
  byGroup = FALSE,
  main = "Number of candidate epitopes per position",
  nbk = 100
)
}
\arguments{
\item{resDF}{a data frame or RAPIMHCI or RAPIMHCII class object}

\item{main}{a cg¿haracter with a plot title}

\item{nbk}{number of x axis breaks}
}
\value{
a ggplot2 object
}
\description{
generate a ggplot2, where in the y axis the amount of alleles bindig to the sequence starting at x position is displayed
}
\examples{
\dontrun{
RepresentedRegions(resDF)
}
}
