% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsubread.R
\name{RsubBuildIndex}
\alias{RsubBuildIndex}
\title{RsubBuildIndex}
\usage{
RsubBuildIndex(hgFile)
}
\description{
The index file will be stored in the same reference file and the path and name
will be masked into the configuration file
}
\details{
Creates the reference index file and update the Configuration file
}
