% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan.R
\name{RunAllMHC}
\alias{RunAllMHC}
\title{RunAllMHC}
\usage{
RunAllMHC(
  seqfile,
  alleleGroup = c("HLA", "BoLA", "Gogo", "H", "H2", "Mamu", "Patr", "SLA", "All"),
  rthParam = 0.5,
  rltParam = 2,
  tParam = -99.9002,
  nCores
)
}
\arguments{
\item{seqfile}{character with the file path of the fasta file (it should be .fasta and contain onle 1 sequence, Multifasta files in progress)}

\item{alleleGroup}{a character vector indicating whi allele family, possible values are "HLA","BoLA", "Gogo","H","H2", "Mamu", "Patr","SLA", "All"}

\item{rthParam}{float (default 0.5) upper limit for strong binder (SB -> peptide percent rank < rhtParam)}

\item{tParam}{tparam}

\item{nCores}{integer indicating the number of used CPUs or workers (see BiocParallele)}

\item{rlhParam}{float (default 2.0) upper limit for weak binder (WB ->  rhtParam <= peptide percent rank < rltParam )}
}
\value{
a character
}
\description{
Run a ALL the available MHCI peptides or by specie
}
\details{
run netMHCIIpan trhough the sequence fasta file for each MHCII Allele
}
