% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan.R
\name{RunAllMHCII}
\alias{RunAllMHCII}
\title{RunMHCIIAlleles}
\usage{
RunAllMHCII(
  seqfile,
  alleleGroup,
  rankS = 0.5,
  rankW = 10,
  rankF = rankW,
  pepLength = 15,
  nCores
)
}
\arguments{
\item{seqfile}{character with the file path of the fasta file (it should be .fasta and contain onle 1 sequence)}

\item{alleleGroup}{see GetAlleleList}

\item{rankS}{float (default 0.5) upper limit for strong binder}

\item{rankW}{float (default 10.0) upper limit for weak binder}

\item{rankF}{float (default 10.0) filter to show peptide binders}

\item{pepLength}{integer (default 15) length of the tested peptide}

\item{nCores}{number of used CPS or worker for BiocParallel}
}
\value{
an RAPIMHCII data.frame like obsejt
}
\description{
Run a peptides between 8 to 14 mers along a fasta sequence (from file)
}
\details{
run ALL MHCII alleles agains the sequenfile
}
