% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS.R
\name{RunFastQC}
\alias{RunFastQC}
\title{R interface to FastQC. It requires that FastQC be installed by means if InstallFastQC function
path location should be saved in the ConfigFile. Try first CheckFastQC().}
\usage{
RunFastQC(
  file,
  paired = TRUE,
  outdir,
  java,
  min_length,
  threads,
  contaminants,
  adapters,
  limits,
  kmers
)
}
\arguments{
\item{file}{a character vector with the file names or the files names to be processed. The files should end in .fastq. If paired-ends, (see paired)
Only the firs file should be indicated (XXX_1.fastq). The second file (XXX_2.fastq) will be automatically searched.}

\item{paired}{(TRUE default) if paired == TRUE, the it will look for the second fastq file for each file. The files are expected to be
.fastq and identified as XXXX_1.fastq and their corresponding pair XXXX_2.fastq.}

\item{outdir}{(character) if missing the output will be stored in a directory named FASQC_RAPI}

\item{java}{(character to  java) Provides the full path to the java binary you want to use to launch fastqc.
If not supplied then java is assumed to be in your path.}

\item{min_length}{Sets an artificial lower limit on the length of the sequence to be shown in the report.}

\item{threads}{( integer) (default 4) number of cores or threads to use (see --cores at TrimGalore web site)}

\item{contaminants}{Specifies a non-default file which contains the list of contaminants to screen overrepresented sequences against.
The file must contain sets of named contaminants in the form name[tab]sequence.  Lines prefixed with a hash will be ignored.}

\item{adapters}{Specifies a non-default file which contains the list of adapter sequences which will be explicity searched against
the library. The file must contain sets of named adapters in the form name[tab]sequence.  Lines prefixed with a hash will be ignored}

\item{limits}{Specifies a non-default file which contains a set of criteria which will be used to determine the warn/error limits for the
various modules.  This file can also be used to selectively remove some modules from the output all together.  The format
needs to mirror the default limits.txt file found in the Configuration folder.}

\item{kmers}{(integer) default 7.  Specifies the length of Kmer to look for in the Kmer content module. Specified Kmer length must be between 2 and 10.}
}
\value{
none
}
\description{
R interface to FastQC. It requires that FastQC be installed by means if InstallFastQC function
path location should be saved in the ConfigFile. Try first CheckFastQC().
}
\details{
the zip and html files for further exploration
}
\author{
Elmer A. Fernández
}
