% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan.R
\name{RunMHCAlleles}
\alias{RunMHCAlleles}
\title{RunMHCAlleles}
\usage{
RunMHCAlleles(
  seqfile,
  alleleList,
  rthParam = 0.5,
  rltParam = 2,
  tParam = -99.9002,
  nCores
)
}
\arguments{
\item{seqfile}{character with the file path of the fasta file (it should be .fasta and contain onle 1 sequence, Multifasta files in progress)}

\item{alleleList}{a list of characters containing alleles list(l1 = c("HLA-A01:01","HLA-A02:01"), l2 = c("HLA-A26:01","HLA-B08:01"))}

\item{rthParam}{float (default 0.5) upper limit for strong binder (SB -> peptide percentage rank < rhtParam)}

\item{tParam}{tparam}

\item{nCores}{integer indicating the number of used CPUs or workers from BiocParallele}

\item{rlhParam}{float (default 2.0) upper limit for weak binder WB ->  rhtParam <= peptide percentrank < rltParam}
}
\value{
a character
}
\description{
Run a ALL the available MHCI peptides or by specie (This function is called by RunAllMHC)
}
\details{
run netMHCIIpan trhough the sequence fasta file for each MHCII Allele
}
