% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan.R
\name{RunMHCIIAlleles}
\alias{RunMHCIIAlleles}
\title{RunMHCIIAlleles}
\usage{
RunMHCIIAlleles(
  seqfile,
  alleleList,
  rankS = 0.5,
  rankW = 10,
  rankF = rankW,
  pepLength = 15,
  nCores
)
}
\arguments{
\item{seqfile}{character with the file path of the fasta file (it should be .fasta and contain onle 1 sequence)}

\item{alleleList}{a character vector with the MHCII allele sequences  c("DRB1_0101") or c("DRB1_0101","DRB1_0111") and so on. If NULL it will be automatically downloaded from the netMHCIIpan server (it may take time)}

\item{rankS}{float (default 0.5) upper limit for strong binder}

\item{rankW}{float (default 10.0) upper limit for weak binder}

\item{rankF}{float (default 10.0) filter to show peptide binders}

\item{pepLength}{integer (default 15) length of the tested peptide}

\item{nCores}{number of used CPS or worker for BiocParallel}
}
\value{
a data.frame
}
\description{
Run a peptides between 8 to 14 mers along a fasta sequence (from file)
}
\details{
run netMHCIIpan trhough the sequence fasta file for each MHCII Allele
}
