% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMHCpan.R
\name{RunNetMHCPan}
\alias{RunNetMHCPan}
\title{RunNetMHCPan}
\usage{
RunNetMHCPan(
  seqfile,
  allele,
  rthParam = 0.5,
  rltParam = 2,
  tParam = -99.9002,
  pLength
)
}
\arguments{
\item{seqfile}{character with the file path of the fasta file (it should be .fasta and contain onle 1 sequence)}

\item{allele}{a character vector with the HLA sequences  c("HLA-A01:01") or c("HLA-A01:01","HLA-A02:01") and so on. If NULL it will be automatically downloaded from the netMHCpan server (it may take time)}

\item{rthParam}{float (default 0.5) upper limit for strong binder (SB -> peptide percent rank < rhtParam)}

\item{tParam}{tparam}

\item{rlhParam}{float (default 2.0) upper limit for weak binder (WB ->  rhtParam <= peptide percent rank < rltParam )}
}
\value{
a character
}
\description{
Run a peptides between 8 to 14 mers along a fasta sequence (from file)
}
\details{
run netMHCpan trhough the sequence fasta file for each HLA
}
