% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS.R
\name{RunTrimGalore}
\alias{RunTrimGalore}
\title{R interface to TrimGalore
please see TrimGalore web site at https://github.com/FelixKrueger/TrimGalore/blob/master/Docs/Trim_Galore_User_Guide.md}
\usage{
RunTrimGalore(
  file,
  paired = TRUE,
  quality = 20,
  phred33 = TRUE,
  phred64 = FALSE,
  fastqc = FALSE,
  max_length,
  error_rate = 0.1,
  gzip = TRUE,
  length = 20,
  maxn,
  trimn,
  outdir,
  report_file = TRUE,
  clip_R1,
  clip_R2,
  three_prime_clip_R1,
  three_prime_clip_R2,
  retain_unpaired,
  length_1 = 35,
  length_2 = 35,
  trim1,
  cores = 4,
  remove_intermediate = paired
)
}
\arguments{
\item{file}{a character vector with the file names or the files names to be processed. The fasts should end in .fastq}

\item{quality}{(integer>0) see --quality at TrimGalore web site}

\item{phred33}{(logical) if TRUE, then --phred33 ON, else --phred33 OFF (see at TrimGalore web site)}

\item{phred64}{(logical) (see --phred64 at TrimGalore web site)}

\item{max_length}{(integer) (see --max_length at TrimGalore web site)}

\item{error_rate}{(float) default 0.1 (see -e option at TrimGalore web site)}

\item{gzip}{(logical) if TRUE (default) the output files will be zipped (see--gzip and --dont_zip at TrimGalore web site)}

\item{length}{(integer) default 20 (see --length at TrimaGalore web site)}

\item{maxn}{(integer) (see --max_n at TrimGalore web site)}

\item{trimn}{(integer) (see --trim-n at TrimGalore web site)}

\item{outdir}{(character) if missing the output will be stored at the same directory of the (first) input file. (see --output_dir at TrimGalore web site)}

\item{report_file}{(logical) (default TRUE) (see --repot_file at TrimGalore web site)}

\item{clip_R1}{(integer) (see --clip_R1 at TrimGalore web site)}

\item{clip_R2}{(integer) (see --clip_R2 at TrimGalore web site)}

\item{three_prime_clip_R1}{(see --three_prime_clip_R1 at TrimGalore web site)}

\item{three_prime_clip_R2}{(see --three_prime_clip_R2 at TrimGalore web site)}

\item{retain_unpaired}{(see --retain_unpaired at TrimGalore web site)}

\item{length_1}{(integer) default 35)  (see --length_1 at TrimGalore web site)}

\item{length_2}{(integer) default 35) (see --length_2 at TrimGalore web site)}

\item{trim1}{(integer) (see --trim1 at TrimGalore web site)}

\item{cores}{( integer) (default 4) number of cores or threads to use (see --cores at TrimGalore web site)}

\item{remove_intermediate}{remove intermediate trimmed files (only fif paired == TRUE)}

\item{pared}{(TRUE default) if paired == TRUE, the it will look for the second fastq file for each file. The files are expected to be
.fastq and identified as XXXX_1.fastq and their corresponding pair XXXX_2.fastq.}

\item{fasqc}{(logical) if TRUE will run fastqc in its default version}
}
\value{
none
}
\description{
R interface to TrimGalore
please see TrimGalore web site at https://github.com/FelixKrueger/TrimGalore/blob/master/Docs/Trim_Galore_User_Guide.md
}
\details{
see output files definitions at TrimGalore web site
}
\author{
Elmer A. Fernández
}
