% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralRAPIFunctions.R
\name{SaveExcel}
\alias{SaveExcel}
\title{SaveExcel}
\usage{
SaveExcel(eDB, file)
}
\arguments{
\item{eDB}{RAPIMHC or RAPIMHCII object}

\item{file}{character string with the file name}
}
\value{

}
\description{
generate a ggplot2, where in the y axis the amount of alleles bindig to the sequence starting at x position is displayed
}
\examples{
\dontrun{
VerifySeq(eDB = result)
}
}
