% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralRAPIFunctions.R
\name{summary.RAPIMHCII}
\alias{summary.RAPIMHCII}
\title{summary}
\usage{
\method{summary}{RAPIMHCII}(eDB)
}
\arguments{
\item{eDB}{RAPIMHC or RAPIMHCII object}
}
\value{

}
\description{
a summary general function to display RAPIMHC and RAPIMHCII results
}
\examples{
\dontrun{
summary(eDB)
}
}
