% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralRAPIFunctions.R
\name{VerifySeq}
\alias{VerifySeq}
\title{VerifySeq}
\usage{
VerifySeq(seqFile, eDB)
}
\arguments{
\item{seqFile}{fasta file name (default missing, it will open an file open dialog box)}

\item{eDB}{RAPIMHC or RAPIMHCII object}
}
\value{
a list with the slot "Peptide" for RAPIMHC y RAPIMHCII and with "InterCore" in the RAPIMHC case
}
\description{
generate a ggplot2, where in the y axis the amount of alleles bindig to the sequence starting at x position is displayed
}
\examples{
\dontrun{
VerifySeq(eDB = result)
}
}
