% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralRAPIFunctions.R
\name{FormatOut}
\alias{FormatOut}
\title{FormatOut}
\usage{
FormatOut(resRAPIMHC)
}
\arguments{
\item{resRAPIMHC}{an object results}
}
\value{

}
\description{
generate a ggplot2, where in the y axis the amount of alleles bindig to the sequence starting at x position is displayed
}
\examples{
\dontrun{
FormatOut(eDB = result)
}
}
