% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAPInetMHCpan.R
\name{installNetMHCPan}
\alias{installNetMHCPan}
\title{Painless installation tool for netMHCpan in linux and R}
\usage{
installNetMHCPan(file = NULL, data = NULL, dir)
}
\arguments{
\item{file}{NULL (default) or character with the netMHCpanXX.Linux.tar.gz file. If NULL, a file selection dialog window will be opened.}

\item{data}{NULL(default) or character with the path of the data.Linux.tar.gz data file. If NULL it will be automatically downloaded from the netMHCpan server (it may take time)}

\item{dir}{character  directory of dir the software will be installed}
}
\value{
if instalation succeed, the following message is printed "netMHCpan Instaltion OK
}
\description{
Painless installation tool for netMHCpan in linux and R
}
\details{
install the netMHCpan software in your computer and allows suing it though R.
 as a results it will create an RDS file named which is used to run the rest of the function
 configIniRAPInetMHCpan.RDS PLEASE DO NOT DELET IT

In this way, it may be run in parallel to speed up peptides searching though the whole HLA set.
Is successful installation , an output displaying the following information should be seen on screen.
}
