% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRAPIFunctions.R
\name{PlotBindingPeptideDistribution}
\alias{PlotBindingPeptideDistribution}
\title{PlotBindingPeptideDistribution}
\usage{
PlotBindingPeptideDistribution(resDF, pepLevel = 0, main)
}
\arguments{
\item{resDF}{a data frame or RAPIMHCI or RAPIMHCII class object}

\item{pepLevel}{peptide level filter ¡?}

\item{main}{a cg¿haracter with a plot title}
}
\value{
a ggplot2 object
}
\description{
generate a ggplot2, where in the y axis the amount of alleles bindig to the sequence starting at x position is displayed
}
\examples{
\dontrun{
RepresentedRegions(resDF)
}
}
